+++ packages/leaflet/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./icons/marker.icon";
-export * from "./map-controls";